pyproject.toml
[tool.poetry]
name = "Zerion Scoring"
version = "0.1.0"
description = ""
authors = ["GGWP"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"
python-socketio = "4.6.1"
aiohttp = "^3.7.4"
asyncio = "^3.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
